Langchain-Practical-Learning/pyproject.toml
[project]
name = "langchain-practical-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-ollama>=1.0.1",
    "pyowm>=3.5.0",
    "python-dotenv>=1.2.1",
]

